+++ hs-bindgen/src-internal/HsBindgen/IR/Hs/Type.hs
@@ -83,4 +83,6 @@ data Type =
   | StrLit String
   | WithFlam Type Type
   | EquivStorable Type
+    -- | 'HsBindgen.Runtime.Struct.IsStructViaReadRaw'
+  | IsStructViaReadRaw Type
   deriving stock (Generic, Show, Eq)