+++ hs-bindgen/test-artefacts/fixtures/macros/macro_in_fundecl_vs_typedef.2.raw/Example.hs
@@ -27,6 +27,7 @@ module Example
 
 import qualified HsBindgen.Runtime.HasCField as HasCField
 import qualified HsBindgen.Runtime.Marshal as Marshal
+import qualified HsBindgen.Runtime.Struct as Struct
 import qualified HsBindgen.Runtime.Support as BG
 import qualified HsBindgen.Runtime.Support.CompatHasField as BG.CompatHasField
 
@@ -125,6 +126,8 @@ instance Marshal.WriteRaw Struct1 where
 
 deriving via Marshal.EquivStorable Struct1 instance BG.Storable Struct1
 
+deriving via Struct.IsStructViaReadRaw Struct1 instance Struct.IsStruct Struct1
+
 {-| __C declaration:__ @a@
 
     __defined at:__ @macros\/macro_in_fundecl_vs_typedef.h 18:30@
@@ -191,6 +194,8 @@ instance Marshal.WriteRaw Struct2 where
 
 deriving via Marshal.EquivStorable Struct2 instance BG.Storable Struct2
 
+deriving via Struct.IsStructViaReadRaw Struct2 instance Struct.IsStruct Struct2
+
 {-| __C declaration:__ @a@
 
     __defined at:__ @macros\/macro_in_fundecl_vs_typedef.h 19:30@
@@ -256,6 +261,8 @@ instance Marshal.WriteRaw Struct3 where
             HasCField.writeRaw (BG.Proxy @"struct3_a") ptr0 struct3_a2
 
 deriving via Marshal.EquivStorable Struct3 instance BG.Storable Struct3
+
+deriving via Struct.IsStructViaReadRaw Struct3 instance Struct.IsStruct Struct3
 
 {-| __C declaration:__ @a@
 
@@ -359,6 +366,8 @@ instance Marshal.WriteRaw Struct4 where
             HasCField.writeRaw (BG.Proxy @"struct4_a") ptr0 struct4_a2
 
 deriving via Marshal.EquivStorable Struct4 instance BG.Storable Struct4
+
+deriving via Struct.IsStructViaReadRaw Struct4 instance Struct.IsStruct Struct4
 
 {-| __C declaration:__ @a@