+++ hs-bindgen/CHANGELOG.md
@@ -179,8 +179,8 @@
   #2060][is-2060] and [PR #2091][pr-2091].
 * Generate `GHC.Records.Compat.HasField` instances for unions. See [issue
   #2060][is-2060] and [PR #2091][pr-2091].
-* Generate `HsBindgen.Runtime.Union.IsUnion` instances for unions and newtypes
-  around unions. See [issue #2060][is-2060] and [PR #2091][pr-2091].
+* Generate `HsBindgen.Runtime.Union.IsUnion` instances for unions. See [issue
+  #2060][is-2060] and [PR #2091][pr-2091].
 * Add `enum` specification to the C type specification of binding
   specifications, enabling users to specify a C `enum` as `closed` so that we
   generate a `COMPLETE` pragma for the declared patterns.
@@ -190,6 +190,8 @@
   Indirect fields are fields of anonymous structs and unions that can be
   accessed as if they were fields of the enclosing struct or union. See [issue
   #2061][is-2061] and [PR #2136][pr-2136].
+* Generate `HsBindgen.Runtime.Structs.IsStruct` instances for structs. See
+  [issue #2121][is-2121] and [PR #2164][pr-2164].
 
 ### Minor changes
 
@@ -329,6 +331,7 @@
 [is-2061]: https://github.com/well-typed/hs-bindgen/issues/2061
 [is-2064]: https://github.com/well-typed/hs-bindgen/issues/2064
 [is-2083]: https://github.com/well-typed/hs-bindgen/issues/2083
+[is-2121]: https://github.com/well-typed/hs-bindgen/issues/2121
 [is-2166]: https://github.com/well-typed/hs-bindgen/issues/2166
 [pr-1862]: https://github.com/well-typed/hs-bindgen/pull/1862
 [pr-1892]: https://github.com/well-typed/hs-bindgen/pull/1892
@@ -348,6 +351,7 @@
 [pr-2136]: https://github.com/well-typed/hs-bindgen/pull/2136
 [pr-2150]: https://github.com/well-typed/hs-bindgen/pull/2150
 [pr-2153]: https://github.com/well-typed/hs-bindgen/pull/2153
+[pr-2164]: https://github.com/well-typed/hs-bindgen/pull/2164
 
 ## 0.1.0-alpha2 -- 2026-03-27